pyproject.toml
[project]
name = "gemma-hackathon"
version = "0.1.0"
description = "MedGemma hackathon HITL demo for trial criteria extraction and grounding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.0",
    "httpx>=0.28.1",
    "kaggle>=1.7.4.5",
    "pypdf>=5.0.0",
    "pdflatex>=0.1.3",
    "wkhtmltopdf>=0.2",
    "markdown>=3.4.0",
    "playwright>=1.40.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-monorepo-plugin>=1.0.0",
    "mkdocs-mermaid2-plugin>=1.0.0",
    "mkdocstrings[python]>=0.24.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "requests>=2.32.5",
    "tenacity>=8.2.0",
    "pytest-asyncio>=1.3.0",
    "aioresponses>=0.7.8",
    "fastapi>=0.128.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "python-dotenv>=1.2.1",
    "diskcache>=5.6.3",
    "platformdirs>=4.5.1",
    "torch>=2.9.1",
    "langchain>=1.2.6",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.6",
    "transformers>=4.57.6",
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "google-cloud-aiplatform>=1.130.0",
    "vertexai>=1.43.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.8",
    "aioresponses>=0.7.8",
    "gsutil>=5.35",
    "mlflow>=3.8.1",
]

[tool.uv.workspace]
members = [
    "components/api-service",
    "components/extraction-service",
    "components/grounding-service",
    "components/data-pipeline",
    "components/evaluation",
    "components/inference",
    "components/model-training",
    "components/shared",
]

[tool.mypy]
exclude = "^(site|components/.*/tests|instructions|notebooks)/"

[[tool.mypy.overrides]]
module = "fastapi"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydantic"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.engine"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlmodel"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pypdf", "pypdf.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "data_pipeline",
    "data_pipeline.download_protocols",
    "extraction_service",
    "extraction_service.pipeline",
    "shared",
    "shared.models",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "datasets",
    "datasets.*",
    "fastmcp",
    "langchain_core",
    "langchain_core.*",
    "langchain_huggingface",
    "langchain_huggingface.*",
    "langgraph",
    "langgraph.*",
    "peft",
    "peft.*",
    "torch",
    "torch.*",
    "transformers",
    "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["components"]
norecursedirs = ["site", "docs", ".venv", "venv", ".uv_cache", "node_modules", ".git"]
markers = [
    "e2e: end-to-end API workflow tests",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore:Inheritance class AiohttpClientSession.*:DeprecationWarning:google\\.genai\\._api_client",
]
pythonpath = [
    "components/api-service/src",
    "components/data-pipeline/src",
    "components/evaluation/src",
    "components/extraction-service/src",
    "components/grounding-service/src",
    "components/inference/src",
    "components/model-training/src",
    "components/shared/src",
]
asyncio_mode = "auto"
addopts = "--cov=api_service --cov=data_pipeline --cov=evaluation --cov=extraction_service --cov=grounding_service --cov=inference --cov=shared --cov-branch --cov-report=term-missing"

[tool.coverage.run]
source = [
    "components/api-service/src",
    "components/data-pipeline/src",
    "components/evaluation/src",
    "components/extraction-service/src",
    "components/grounding-service/src",
    "components/inference/src",
    "components/shared/src",
]
omit = ["*/tests/*", "*/.venv/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
